Organise classes into view subfolders
